# Fix the rollup.config.js wasm-opt args