--- a/crates/web-client/rollup.config.js
+++ b/crates/web-client/rollup.config.js
@@ -5,6 +5,11 @@ import copy from "rollup-plugin-copy";
 
 // Flag that indicates if the build is meant for testing purposes.
 const testing = process.env.MIDEN_WEB_TESTING === "true";
+const wasmOptArgs = [
+  "-O4",
+  "--enable-bulk-memory",
+  "--enable-nontrapping-float-to-int",
+];
 
 /**
  * Rollup configuration file for building a Cargo project and creating a WebAssembly (WASM) module,
@@ -48,13 +53,17 @@ export default [
             `build.rustflags=["-C", "target-feature=+atomics,+bulk-memory,+mutable-globals", "-C", "link-arg=--max-memory=4294967296"]`,
             "--no-default-features",
           ],
-          wasmOpt: testing
-            ? ["-O0", "--enable-threads", "--enable-bulk-memory-opt"]
-            : ["--enable-threads", "--enable-bulk-memory-opt"],
+          // We need to pass an invalid flag to silently skip the wasm optimization
+          // The correct way of doing this is to build the wasm with the optimize: { release: false }
+          // but building the wasm with the dev profile fails due to a too many locals error
+          // And the plugin doesn't allow us to directly skip the wasm-opt step or specify a non-release/dev profile
+          // We should change this when this issue is fixed: https://github.com/wasm-tool/rollup-plugin-rust/issues/50
+          ...(testing ? { wasmOpt: ["--an-invalid-option"]} : { wasmOpt: wasmOptArgs }),
         },
         experimental: {
           typescriptDeclarationDir: "dist/crates",
         },
+        optimize: { release: true, rustc: true },
       }),
       resolve(),
       commonjs(),