The issue with wasmOpt has been resolved by removing the invalid option `--enable-bulk-memory-opt`. This option caused Wasmopt to fail silently with a warning, resulting in a successful build despite the failure.

To resolve this issue, the solution was to remove the invalid option, allowing WasmOpt to function correctly. 
Although this change increases the build time by approximately 45 seconds to a minute, it results in a significant improvement in performance. 

In testing, this change yielded around a 25% faster improvement for creating new wallets.

I tested it with this branch: https://github.com/demox-labs/miden-client/tree/investigate/account-slow 
For the wasm call to `createWallet` for a private immutable wallet without a seed:

Before:
```
Average: 7313.53ms
Min: 765.74ms
Max: 33621.57ms
All timings: 10912.93, 12244.78, 1240.48, 787.17, 6073.67, 6873.93, 33621.57, 4326.34, 2115.14, 11700.06, 6459.99, 2962.37, 7196.78, 765.74, 5637.33, 3007.66, 2522.94, 9537.67, 20903.81, 1208.49, 5120.15, 2987.92, 11692.99, 6880.08, 6058.19
```

After:
```
Average: 5254.26ms
Min: 758.00ms
Max: 16598.55ms
All timings: 1233.97, 9275.52, 5779.14, 2297.96, 1168.27, 758.00, 6212.18, 768.31, 2421.03, 7895.36, 4496.00, 8637.89, 10762.14, 1597.08, 4943.79, 5783.90, 9524.69, 9044.50, 2862.70, 7428.73, 4969.38, 759.30, 4510.98, 1627.14, 16598.55
```