@evanmarshall 2025-05-30
It's related to this issue: https://github.com/0xMiden/miden-client/issues/892