@evanmarshall 2025-05-31
Yes, it's speeding up build times because it fails quickly with the error: `(!) Plugin rust: wasm-opt failed: Command `wasm-opt index_bg.wasm --output wasm_opt.wasm --enable-threads --enable-bulk-memory-opt` failed with error code: 1`. It's not an available option: https://manpages.debian.org/testing/binaryen/wasm-opt.1.en.html#enable~6 . It works because only it emits a warning.

You can see it catch the error here: https://github.com/wasm-tool/rollup-plugin-rust/blob/master/src/wasm-opt.js#L22 and emit it here: https://github.com/wasm-tool/rollup-plugin-rust/blob/master/src/index.js#L301

I'll address the PR feedback shortly.