@evanmarshall 2025-06-01
Sorry about that, I was building locally with nightly. I made sure it worked with stable.