@igamigo 2025-06-01
I ran integration tests locally from your branch and they are passing. I think you might have to rebase off of the latest `next`. I think we want to include this in a patch release so if you want to rebase it off of `main` (so that we can merge to `main`) that would work too.