@evanmarshall 2025-06-05
I merged in the new wallet speed ups. Test time is down ~9 minutes going from 27min to 16min. 
I think this should be good to go.