@evanmarshall 2025-06-05
About to merge. I struggled with separately the nightly vs stable for different crates. What I was doing worked locally but I was playing guess and check with the CI so I decided stable for now with the wasm. The difference was ~1.2MB in size for the wasm.

I also reduced `O4` to `O3` in the `wasm-opt` settings.